+++ R/convenience-functions.R
@@ -1306,9 +1306,11 @@ exp_pars <- function(formula, data, C) {
     MCM <- M %*% C %*% M
     eigens <- eigen(MCM, symmetric = TRUE)
     npos <- sum(eigens$values > 0)
-    sa <- mc(residuals(lm(formula, data = data)), C)
+    res <- residuals(lm(formula, data = data))
+    obs_idx <- which(!is.na(res))
+    sa <- mc(res[obs_idx], C[obs_idx, obs_idx])
     X <- model.matrix(formula, data)
-    E_sa <- expected_mc(X, C)
+    E_sa <- expected_mc(X, C[obs_idx, obs_idx])
     Sigma_sa <- sqrt( 2 / nlinks )
     z_sa <- (sa - E_sa) / Sigma_sa
     if (z_sa < -.59) {